+++ controllers/client/home.controller.js
@@ -1,0 +1,7 @@
+const getHome = (req, res) => {
+  res.render('client/pages/home.pug', {
+    pageTitle: 'Trang chủ',
+  });
+};
+
+module.exports = { getHome };
