+++ package.json
@@ -1,21 +1,18 @@
 {
-  "name": "hit-nodejs-2025",
+  "name": "bai4",
   "version": "1.0.0",
-  "main": "src/index.js",
+  "main": "index.js",
+  "license": "MIT",
   "scripts": {
-    "dev": "nodemon src/index.js",
-    "start": "node src/index.js",
-    "format": "prettier --write \"src/**/*.js\""
+    "start": "node --watch index.js",
+    "dev": "nodemon --inspect index.js",
+    "format": "prettier --write \"**/*.js\""
   },
-  "keywords": [],
-  "author": "",
-  "license": "ISC",
-  "description": "",
   "dependencies": {
     "dotenv": "^16.4.7",
-    "ejs": "^3.1.10",
     "express": "^4.21.2",
     "http-status-codes": "^2.3.0",
+    "pug": "^3.0.3",
     "uuid": "^11.1.0"
   },
   "devDependencies": {
