+++ routes/client/home.route.js
@@ -1,0 +1,7 @@
+const router = require('express').Router();
+
+const homeController = require('../../controllers/client/home.controller');
+
+router.get('/', homeController.getHome);
+
+module.exports = router;
