--- src/routes/home.route.js
@@ -1,9 +1,0 @@
-const express = require('express');
-
-const homeController = require('../controllers/home.controller');
-
-const router = express.Router();
-
-router.get('/', homeController.homePage);
-
-module.exports = router;
