--- src/routes/user.route.js
@@ -1,12 +1,0 @@
-const express = require('express');
-const router = express.Router();
-
-const userController = require('../controllers/user.controller');
-
-router.get('/', userController.getUsers);
-
-router.get('/:id', userController.getUserById);
-
-router.post('/', userController.createUser);
-
-module.exports = router;
