+++ views/client/layouts/default.pug
@@ -1,0 +1,15 @@
+doctype html
+html(lang="en")
+  head
+    meta(charset="UTF-8")
+    meta(name="viewport", content="width=device-width, initial-scale=1.0")
+    title #{pageTitle}
+    link(rel="stylesheet", href="assets/css/styles.css")
+  body 
+    include ../partials/header.pug
+
+    block main
+    
+    include ../partials/footer.pug
+
+    script(src="assets/js/index.js")
