+++ views/client/pages/home.pug
@@ -1,0 +1,4 @@
+extends ../layouts/default.pug
+
+block main
+  h1 Trang chủ
