+++ views/client/pages/user.pug
@@ -1,0 +1,4 @@
+extends ../layouts/default.pug
+
+block main
+  h1 Trang user
