+++ views/client/partials/footer.pug
@@ -1,0 +1,1 @@
+footer Footer
