+++ views/client/partials/header.pug
@@ -1,0 +1,1 @@
+header Header
