--- views/home/index.ejs
@@ -1,11 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-    <head>
-        <meta charset="UTF-8" />
-        <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-        <title><%= title %></title>
-    </head>
-    <body>
-        <%= message %>
-    </body>
-</html>
